requirements.txt
Django==1.8
django-registration-redux==1.1
python-dateutil==2.4.2
six==1.9.0
wsgiref==0.1.2